requirements.txt
# Core
numpy>=1.21.0
pandas>=1.3.0
scikit-learn>=1.0.0

# NLP
konlpy>=0.6.0
sentence-transformers>=2.2.0
transformers>=4.30.0
torch>=2.0.0

# Topic Modeling
bertopic>=0.15.0
hdbscan>=0.8.29
umap-learn>=0.5.3

# Visualization
matplotlib>=3.5.0
seaborn>=0.12.0
plotly>=5.14.0

# Utils
tqdm>=4.65.0
pyyaml>=6.0
python-mecab-ko>=1.0.14